Move message encoding/decoding
